requirements.txt
anyio==3.5.0
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cookiecutter==1.7.3
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
manage-fastapi==1.1.1
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.0
poyo==0.5.0
prompt-toolkit==3.0.29
pydantic==1.9.0
python-dateutil==2.8.2
python-multipart==0.0.5
python-slugify==6.1.1
questionary==1.10.0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
text-unidecode==1.3
torch==1.11.0+cu115
torchvision==0.12.0+cu115
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
